UI/requirements.txt
allure-pytest==2.13.3
allure-python-commons==2.13.3
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
execnet==2.0.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
py==1.11.0
pycparser==2.21
PySocks==1.7.1
pytest==7.1.2
pytest-rerunfailures==13.0
pytest-xdist==3.5.0
requests==2.31.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.23.2
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
